fix test

+++ pandas/tests/indexing/test_loc.py
@@ -3352,10 +3352,10 @@ class TestLocSeries:
         # GH 57027
         ids = list(range(11))
         index = Index(ids * 1000, dtype="Int64")
-        df = DataFrame({"val": np.arange(len(index))}, index=index)
+        df = DataFrame({"val": np.arange(len(index), dtype=np.intp)}, index=index)
         result = df.loc[ids]
         expected = DataFrame(
-            {"val": index.argsort(kind="stable")},
+            {"val": index.argsort(kind="stable").astype(np.intp)},
             index=Index(np.array(ids).repeat(1000), dtype="Int64"),
         )
         tm.assert_frame_equal(result, expected)
